+++ R/checkVarNames.R
@@ -8,14 +8,16 @@
 #'
 #' Invalid column names in a \code{SQLite} data base include
 #' \itemize{
-#' \item \code{SQLite} keywords (see \code{\link[eatDB]{sqlite_keywords}}) and
-#' \item column names with a \code{"."} in it.
+#' \item \code{SQLite} keywords (see \code{\link[eatDB]{sqlite_keywords}}),
+#' \item column names with a \code{"."} in it and
+#' \item duplicate variable names which arise from \code{SQLite} being case insensitive.
 #' }
 #'
 #' The corresponding variable name changes are
 #' \itemize{
-#' \item appending the suffix \code{"Var"} to all \code{SQLite} keywords and
-#' \item changing all \code{"."} in variable names to \code{"_"}.
+#' \item appending the suffix \code{"Var"} to all \code{SQLite} keywords,
+#' \item changing all \code{"."} in variable names to \code{"_"} and
+#' \item appending \code{"_2"} to case insensitive duplicated variable names.
 #' }
 #'
 #'Note that avoiding \code{"."} in variable names is beneficial for multiple reasons, such as
@@ -24,6 +26,7 @@
 #'@param GADSdat \code{GADSdat} or \code{all_GADSdat} object.
 #'@param checkKeywords Logical. Should \code{SQLite} keywords be checked and modified?
 #'@param checkDots Logical. Should occurrences of \code{"."} be checked and modified?
+#'@param checkDuplicates Logical. Should case insensitive duplicate variable names be checked and modified?
 #'
 #'@return Returns the original object with updated variable names.
 #'
@@ -35,18 +38,18 @@
 #' pisa3 <- checkVarNames(pisa2)
 #'
 #'@export
-checkVarNames <- function(GADSdat, checkKeywords = TRUE, checkDots = TRUE) {
+checkVarNames <- function(GADSdat, checkKeywords = TRUE, checkDots = TRUE, checkDuplicates = TRUE) {
   UseMethod("checkVarNames")
 }
 #'@export
-checkVarNames.GADSdat <- function(GADSdat, checkKeywords = TRUE, checkDots = TRUE) {
+checkVarNames.GADSdat <- function(GADSdat, checkKeywords = TRUE, checkDots = TRUE, checkDuplicates = TRUE) {
   check_GADSdat(GADSdat)
   GADSdat[["labels"]][, "varName"] <- sapply(GADSdat[["labels"]][, "varName"], checkVarNames)
   names(GADSdat[["dat"]]) <- sapply(names(GADSdat[["dat"]]), checkVarNames)
   GADSdat
 }
 #'@export
-checkVarNames.all_GADSdat <- function(GADSdat, checkKeywords = TRUE, checkDots = TRUE) {
+checkVarNames.all_GADSdat <- function(GADSdat, checkKeywords = TRUE, checkDots = TRUE, checkDuplicates = TRUE) {
   check_all_GADSdat(GADSdat)
   GADSdat[["allLabels"]][, "varName"] <- sapply(GADSdat[["allLabels"]][, "varName"], checkVarNames)
   GADSdat[["datList"]] <- lapply(GADSdat[["datList"]], function(df) {
@@ -56,12 +59,12 @@ checkVarNames.all_GADSdat <- function(GADSdat, checkKeywords = TRUE, checkDots =
   GADSdat
 }
 #'@export
-checkVarNames.data.frame <- function(GADSdat, checkKeywords = TRUE, checkDots = TRUE) {
+checkVarNames.data.frame <- function(GADSdat, checkKeywords = TRUE, checkDots = TRUE, checkDuplicates = TRUE) {
   names(GADSdat) <- checkVarNames(names(GADSdat))
   GADSdat
 }
 #'@export
-checkVarNames.character <- function(GADSdat, checkKeywords = TRUE, checkDots = TRUE) {
+checkVarNames.character <- function(GADSdat, checkKeywords = TRUE, checkDots = TRUE, checkDuplicates = TRUE) {
   NewName <- GADSdat
   check_logicalArgument(checkKeywords)
   check_logicalArgument(checkDots)
@@ -80,6 +83,12 @@ checkVarNames.character <- function(GADSdat, checkKeywords = TRUE, checkDots = T
   if(checkDots){
     NewName <- gsub("\\.", "_", NewName)
   }
+  ## Check for duplicates due to SQLite ignoring case
+  if(checkDuplicates) {
+    if(anyDuplicated(tolower(NewName))) {
+      NewName <- unduplicate(NewName)
+    }
+  }
 
   ## report all changes
   which_changed <- which(NewName != GADSdat)
@@ -89,4 +98,14 @@ checkVarNames.character <- function(GADSdat, checkKeywords = TRUE, checkDots = T
     }
   }
   NewName
+}
+
+# sqlite3 not case sensitive!
+unduplicate <- function(x) {
+  out <- x
+  allower <- tolower(x)
+  out[duplicated(allower)] <- paste(out[duplicated(allower)], 2, sep = "_")
+  if(anyDuplicated(tolower(out))) out <- unduplicate(out)
+
+  out
 }